:adhesive_bandage: rename `u254` to `bigint_mul`
