+++ bitcoin-testscripts/src/friendly/bigint_mul.rs
