+++ bitcoin-testscripts/src/friendly/mod.rs
@@ -1,1 +1,1 @@
-pub mod u254;
+pub mod bigint_mul;
